Create Flappy Bird Game in Python with Pygame
This Python-based project, "Flappy Bird Game," is implemented using the Pygame library. It's a classic and addictive game where players control a bird that must navigate through a series of pipes without colliding with them or hitting the ground. The game features smooth graphics, dynamic animation, a score system, and a game-over screen. Test your reflexes and see how far you can go in this challenging and entertaining game!